CloudFunction/requirement.txt
keras==2.4.3
matplotlib==3.4.2
numpy
protobuf==3.17.1
tensorflow==2.4.1
google-cloud-storage==1.16.1
Pillow==6.2.0

# Function dependencies, for example:
# package>=version